fix: if the popup was not opened, the extension was always disabled
Because the popup was not opened every time we navigate on a page (and that's normal), the inject script was not able to update the extension status on the page.

Now, when the injected script finishes loading, it sends a message to the background to ask for the extension status.
The background then get from the storage the status and communicates it to the injected script.

The injected script continues what it should do but with the corrected value.